backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.23
sqlite3

# Vector DB & Embeddings
chromadb==0.4.14
sentence-transformers==2.2.2
numpy==1.24.3

# Google Cloud & ADK
google-cloud-storage==2.14.0
google-cloud-vision==3.4.0
google-generativeai==0.3.0
google-adk
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ML & NLP
PyPDF2==3.0.1
python-docx==0.8.11
pydantic-ai==0.0.3

# Authentication & Security
PyJWT==2.8.1
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
cryptography==41.0.7

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging & Monitoring
python-json-logger==2.0.7
